Fixed for Ubuntu 18.04+ and Updated Versions
Fixed to work with version of Ubuntu from 16.04 and prior and 16.10 and later. Had to compile libtorrent and rtorrent from source for this (error in rakshasa's master repository). 

Also updated all the software versions for additional software that is downloaded. Fully tested with all 3 webservers in Ubuntu 18.04 and tested with lighttpd on 16.04 (but all 3 webservers should work).